Audit/requirements.txt
a2wsgi==1.10.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
connexion==3.0.5
Flask==3.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kazoo==2.5.0
MarkupSafe==2.1.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.1
pykafka==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
swagger_ui_bundle==1.1.0
tabulate==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
